Add files via upload
The purpose of this assignment is to implement a simple workflow that will assess the similarity between supermarket customers and suggest for any input customer a list of his/her 10 most similar other customers. In order to calculate the similarity between customers we will first have to compute the dissimilarity for every given attribute.